API switch to support both older and newer versions of anki
